fix: try to componentize intersection observer functionality
